requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
python-telegram-bot==20.7
schedule==1.2.1
python-dotenv==1.0.0
pandas==2.1.4
lxml==4.9.3 